requirements.txt
django==4.2.5
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2==2.9.8
drf-nested-routers==0.93.4